pyproject.toml
[project]
name = "alpakabooking"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "mypy>=1.15.0",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "pytz>=2025.1",
    "pyyaml>=6.0.2",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "djangorestframework-simplejwt>=5.5.0",
    "django-cors-headers>=4.7.0",
    "drf-spectacular>=0.28.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "rosenhof.settings"
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "fe",
    "migrations",
]

line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"